requirements.txt
# packages in environment at /home/songhongwei/anaconda2/envs/asc_mt:
#
audioread                 2.1.5                    py36_0    conda-forge
backcall                  0.1.0                    py36_0  
blas                      1.0                         mkl  
bleach                    3.0.2                    py36_0  
bzip2                     1.0.6                h470a237_2    conda-forge
ca-certificates           2018.03.07                    0  
certifi                   2018.10.15               py36_0  
cffi                      1.11.5           py36he75722e_1  
cycler                    0.10.0                     py_1    conda-forge
dbus                      1.13.2               h714fa37_1  
decorator                 4.3.0                      py_0    conda-forge
entrypoints               0.2.3                    py36_2  
expat                     2.2.5                hfc679d8_2    conda-forge
ffmpeg                    2.8.6                hf82bc7d_7    conda-forge
fontconfig                2.13.1               h65d0f4c_0    conda-forge
freetype                  2.9.1                h8a8886c_1  
gettext                   0.19.8.1             h5e8e0c9_1    conda-forge
glib                      2.56.2               h464dc38_1    conda-forge
gmp                       6.1.2                h6c8ec71_1  
gst-plugins-base          1.14.0               hbbd80ab_1  
gstreamer                 1.14.0               hb453b48_1  
h5py                      2.8.0            py36h61e79e4_3  
hdf5                      1.8.20               hba1933b_1  
icu                       58.2                 hfc679d8_0    conda-forge
intel-openmp              2019.0                      118  
ipykernel                 5.1.0            py36h39e3cac_0  
ipython                   7.1.1            py36h39e3cac_0  
ipython_genutils          0.2.0                    py36_0  
ipywidgets                7.4.2                    py36_0  
jedi                      0.13.1                   py36_0  
jinja2                    2.10                     py36_0  
joblib                    0.13.0                     py_0    conda-forge
jpeg                      9b                   h024ee3a_2  
jsonschema                2.6.0                    py36_0  
jupyter                   1.0.0                    py36_7  
jupyter_client            5.2.3                    py36_0  
jupyter_console           6.0.0                    py36_0  
jupyter_core              4.4.0                    py36_0  
kiwisolver                1.0.1            py36h2d50403_2    conda-forge
libedit                   3.1.20170329         h6b74fdf_2  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 8.2.0                hdf63c60_1  
libgfortran-ng            7.3.0                hdf63c60_0  
libiconv                  1.15                 h470a237_3    conda-forge
libpng                    1.6.35               hbc83047_0  
librosa                   0.6.2                      py_0    conda-forge
libsodium                 1.0.16               h1bed415_0  
libstdcxx-ng              8.2.0                hdf63c60_1  
libtiff                   4.0.9                he85c1e1_2  
libuuid                   2.32.1               h470a237_2    conda-forge
libxcb                    1.13                 h470a237_2    conda-forge
libxml2                   2.9.8                h422b904_5    conda-forge
llvmlite                  0.25.0           py36hd28b015_1    conda-forge
markupsafe                1.0              py36h14c3975_1  
matplotlib                3.0.1                h8a2030e_1    conda-forge
matplotlib-base           3.0.1            py36hc039c98_1    conda-forge
metric-learn              0.4.0                     <pip>
mistune                   0.8.4            py36h7b6447c_0  
mkl                       2019.0                      118  
mkl_fft                   1.0.6            py36h7dd41cf_0  
mkl_random                1.0.1            py36h4414c95_1  
nbconvert                 5.3.1                    py36_0  
nbformat                  4.4.0                    py36_0  
ncurses                   6.1                  hf484d3e_0  
ninja                     1.8.2            py36h6bb024c_1  
notebook                  5.7.0                    py36_0  
numba                     0.40.0           py36hf8a1672_0    conda-forge
numpy                     1.15.4           py36h1d66e8a_0  
numpy-base                1.15.4           py36h81de0dd_0  
olefile                   0.46                     py36_0  
openssl                   1.0.2p               h470a237_1    conda-forge
pandoc                    2.2.3.2                       0  
pandocfilters             1.4.2                    py36_1  
parso                     0.3.1                    py36_0  
pcre                      8.42                 h439df22_0  
pexpect                   4.6.0                    py36_0  
pickleshare               0.7.5                    py36_0  
pillow                    5.3.0            py36h34e0f95_0  
pip                       18.1                     py36_0  
prometheus_client         0.4.2                    py36_0  
prompt_toolkit            2.0.7                    py36_0  
pthread-stubs             0.4                  h470a237_1    conda-forge
ptyprocess                0.6.0                    py36_0  
pycparser                 2.19                     py36_0  
pygments                  2.2.0                    py36_0  
pyparsing                 2.3.0                      py_0    conda-forge
pyqt                      5.6.0            py36h8210e8a_7    conda-forge
python                    3.6.6                h5001a0f_3    conda-forge
python-dateutil           2.7.5                      py_0    conda-forge
pytorch                   0.4.1           py36_py35_py27__9.0.176_7.1.2_2    pytorch
pyzmq                     17.1.2           py36h14c3975_0  
qt                        5.6.3                h39df351_1  
qtconsole                 4.4.2                    py36_0  
readline                  7.0                  h7b6447c_5  
resampy                   0.2.1                      py_1    conda-forge
scikit-learn              0.20.0           py36h4989274_1  
scipy                     1.1.0            py36hfa4b5c9_1  
send2trash                1.5.0                    py36_0  
setuptools                40.5.0                   py36_0  
sip                       4.18.1           py36hfc679d8_0    conda-forge
six                       1.11.0                   py36_1  
sqlite                    3.25.2               h7b6447c_0  
terminado                 0.8.1                    py36_1  
testpath                  0.4.2                    py36_0  
tk                        8.6.8                hbc83047_0  
torchsummary              1.5.1                     <pip>
torchvision               0.2.1                    py36_1    pytorch
tornado                   5.1.1            py36h470a237_0    conda-forge
tqdm                      4.28.1           py36h28b3542_0  
traitlets                 4.3.2                    py36_0  
wcwidth                   0.1.7                    py36_0  
webencodings              0.5.1                    py36_1  
wheel                     0.32.2                   py36_0  
widgetsnbextension        3.4.2                    py36_0  
x264                      1!152.20180717       h470a237_1    conda-forge
xgboost                   0.81                      <pip>
xorg-libxau               1.0.8                h470a237_6    conda-forge
xorg-libxdmcp             1.1.2                h470a237_7    conda-forge
xz                        5.2.4                h14c3975_4  
zeromq                    4.2.5                hf484d3e_1  
zlib                      1.2.11               ha838bed_2  
